% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_App.R
\name{convertToNum}
\alias{convertToNum}
\title{Conversion of expression to numbers}
\usage{
convertToNum(l)
}
\arguments{
\item{l}{is an expression which should be evaluated. If the expression can
be evaluated to a number the number is returned. Otherwise an error is returned.}
}
\description{
converts an expression to a number
}
