% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePolynome.R
\name{createPolynom}
\alias{createPolynom}
\title{Converts an algebraic system to a polynome by eliminating variables}
\usage{
createPolynom(f, elimVars)
}
\arguments{
\item{f}{is a function defining the algebraic system}

\item{elimVars}{is a character vector defining the variables which have to be eliminated}
}
\value{
the resulting polynome of the elimination
}
\description{
Converts an algebraic system to a polynome by eliminating variables
}
\examples{
f <- function() {
  h + hd + -h0 = 0
  d + hd -d0 = 0   
  hd / (h*d) -kd = 0
}
elimVars <- c("h", "d")
createPolynom(f, elimVars)
}
