% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionUtils.R
\name{ErrorClass}
\alias{ErrorClass}
\title{ErrorClass class for handling errors}
\description{
a class for handling error messages
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{message}}{the error message}

\item{\code{object}}{an R object which can be stored in the class instance if an error and a result should be returned together.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ErrorClass-new}{\code{ErrorClass$new()}}
\item \href{#method-ErrorClass-clone}{\code{ErrorClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErrorClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-ErrorClass-new}{}}}
\subsection{Method \code{new()}}{
create a new ErrorClass Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErrorClass$new(message, object = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{the string describing the error}

\item{\code{object}}{is optional if something besides the message should be stored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ErrorClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ErrorClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ErrorClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
