% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{opti}
\alias{opti}
\title{Optimize algebraic systems which describe thermodynamic binding systems}
\usage{
opti(
  case,
  lowerBounds,
  upperBounds,
  path,
  additionalParameters,
  seed = NULL,
  npop = 40,
  ngen = 200,
  Topology = "random",
  errorThreshold = -Inf,
  error_calc_fct = "Rel. Error",
  add_info = "",
  engine = c("ast2ast", "r")
)
}
\arguments{
\item{case}{is a character describing which system should be investigated. Either:
"dba_host_const",
"dba_dye_const", "ida" or "gda".}

\item{lowerBounds}{is a numeric vector defining the lower boundaries of the parameter.
In case of \emph{dba_dye_const} or *dba_host_const the first parameter is: \emph{khd}.
In case of \emph{ida} and \emph{ga} the first parameter is: \emph{kg}.
Afterwards the parameters \emph{I0}, \emph{IHD} and \emph{ID} have to be repeated for each signal respectivly.}

\item{upperBounds}{is a numeric vector defining the upper boundaries of the parameter.
The order is the same as for the lower boundaries.}

\item{path}{is a filepath which contains tabular x-y data.
The concentraion of dye or guest respectivly is assumed to be in the first column.
Furthermore, should the corresponding signals be stored in the other columns.
As an alternative an already loaded data.frame can be passed to the function.}

\item{additionalParameters}{are required parameters which are specific for each case.
In case of \emph{dba_host_const} a numeric vector of length 1 is expected which contains the concentration of the host.
In case of \emph{dba_dye_const} a numeric vector of length 1 is expected which contains the concentration of the dye.
In case of \emph{ida} a numeric vector of length 3 is expected which contains the concentration of the host, dye and the \emph{khd} parameter.
In case of \emph{gda} a numeric vector of length 3 is expected which contains the concentration of the host, guest and the \emph{khd} parameter.}

\item{seed}{is an optional integer argument defining the seed which is set directly for the optimization. In case the argument is not set the current time is used as seed.}

\item{npop}{is an optional integer argument defining the number of particles during optimization. The default value is set to 40.}

\item{ngen}{is an optional integer argument defining the number of generations of the particle swarm optimization. The default value is set to 200.}

\item{Topology}{is an optional character argument defining which topology should be used by the particle swarm algorithm. The options are "star" and "random". The default topology is the "random" topology.}

\item{errorThreshold}{is an optional numeric argument defining a sufficient small error which acts as a stop signal for the particle swarm algorithm. The default value is set to -Inf.}

\item{error_calc_fct}{is an optional character argument defining how the error between the
in silico signal and the measured signal is calculated. One of \emph{Rel. Error}, \emph{RMSE},
\emph{SSE}, or \emph{Huber}. The default is \emph{Rel. Error}.}

\item{add_info}{is an optional character argument which is printed during optimization}

\item{engine}{is an optional character argument selecting the fitness-evaluation backend for
the particle swarm optimization: "ast2ast" (the default) is an ast2ast-compiled
per-particle loss, "r" is the plain-R loss function -- see \code{\link{pso}}'s engine
parameter for details. Results match to floating-point precision either way.}
}
\value{
either an instance of ErrorClass if something went wrong. Otherwise the optimized parameter and the \emph{insilico} signal values are returned.
}
\description{
Optimize algebraic systems which describe thermodynamic binding systems
}
\examples{
path <- paste0(system.file("examples", package = "tsf"), "/IDA.txt")
opti("ida", c(1, 0, 0, 0), c(10^9, 10^6, 10^6, 10^6), path, c(4.3, 6.0, 7079458))
}
