% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizeVapro.R
\name{opti_vapro}
\alias{opti_vapro}
\title{Optimize algebraic systems which describe thermodynamic binding systems using
the VAPRO (variable projection) approach}
\usage{
opti_vapro(
  case,
  lowerBounds,
  upperBounds,
  path,
  additionalParameters,
  nGrid = 1000L,
  error_calc_fct = "Rel. Error"
)
}
\arguments{
\item{case}{is a character describing which system should be investigated. Either:
"dba_host_const",
"dba_dye_const", "ida" or "gda".}

\item{lowerBounds}{is a numeric vector of length 1 defining the lower boundary of the
nonlinear binding parameter. In case of \emph{dba_dye_const} or \emph{dba_host_const}
this is \emph{khd}. In case of \emph{ida} and \emph{gda} this is \emph{kg}.}

\item{upperBounds}{is a numeric vector of length 1 defining the upper boundary of the
nonlinear binding parameter. The order is the same as for the lower boundary.}

\item{path}{is a filepath which contains tabular x-y data.
The concentraion of dye or guest respectivly is assumed to be in the first column.
Furthermore, should the corresponding signals be stored in the other columns.
As an alternative an already loaded data.frame can be passed to the function.}

\item{additionalParameters}{are required parameters which are specific for each case.
In case of \emph{dba_host_const} a numeric vector of length 1 is expected which contains the concentration of the host.
In case of \emph{dba_dye_const} a numeric vector of length 1 is expected which contains the concentration of the dye.
In case of \emph{ida} a numeric vector of length 3 is expected which contains the concentration of the host, dye and the \emph{khd} parameter.
In case of \emph{gda} a numeric vector of length 3 is expected which contains the concentration of the host, guest and the \emph{khd} parameter.}

\item{nGrid}{is an optional integer argument defining the number of exponentially spaced grid points used for the linear search of the nonlinear binding parameter between lowerBounds and upperBounds. The default value is set to 1000.}

\item{error_calc_fct}{is an optional input defining how the error between the in silico signal and the measured signal is calculated.
One can use one of the following predefined functions as character vectors: \emph{Rel. Error}, \emph{RMSE}, \emph{SSE}, or \emph{Huber}. The default function is \emph{Rel. Error}.
Alternatively a function can be passed to opti_vapro, which has to expect two arguments, first the insilico signal followed by the measured signal.}
}
\value{
either an instance of ErrorClass if something went wrong. Otherwise the optimized parameter and the \emph{insilico} signal values are returned.
}
\description{
Compared to \code{\link{opti}} only the nonlinear binding parameter
(Ka(HD) or Ka(HG)) is searched, via a linear grid search over the log
(i.e. exponentially spaced) parameter space, followed by a local
continuous refinement (\code{\link[stats]{optimize}}) restricted to the
bracket around the best grid point. The coarse grid keeps the search
robust against multiple local optima over the wide parameter range, while
the local refinement avoids the discretization artifacts a pure grid
search would otherwise introduce (identical estimates across many
bootstrap replicates whenever the true parameter uncertainty is smaller
than the grid spacing). The linear intercept/slope parameters (I0, IHD,
ID) for each signal are not part of the search space.
Instead they are profiled out in closed form for every candidate binding
parameter using non-negative least squares (\code{\link[nnls]{nnls}}).
}
\examples{
path <- paste0(system.file("examples", package = "tsf"), "/IDA.txt")
opti_vapro("ida", 1, 10^9, path, c(4.3, 6.0, 7079458))
}
