% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Offers a GUI for the optimization of
algebraic systems describing thermodynamic binding systems}
\usage{
runApp(port)
}
\arguments{
\item{port}{is a number defining the port to use.}
}
\description{
Offers a GUI for the optimization of
algebraic systems describing thermodynamic binding systems
}
\examples{
\donttest{
tsf::runApp()
}
}
