% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunBatch.R
\name{batch}
\alias{batch}
\title{Runs a batch of optimization tasks}
\usage{
batch(
  case,
  lowerBounds,
  upperBounds,
  path,
  additionalParameters,
  seed = NA,
  npop = 40,
  ngen = 200,
  Topology = "random",
  errorThreshold = -Inf,
  num_rep = 1,
  num_cores = 1
)
}
\arguments{
\item{case}{is a character argument which specifies the optimization case.
Either "dba_dye_const", "dba_host_const", "ida" or "gda"}

\item{lowerBounds}{is a numeric vector with the lower bounds for the optimization}

\item{upperBounds}{is a numeric vector with the upper bounds for the optimization}

\item{path}{is a character argument which specifies the path to the data}

\item{additionalParameters}{is a numeric vector with additional parameters
In case of \emph{dba_dye_const} or *dba_host_const the order of the parameters is: \emph{khd}, \emph{I0}, \emph{IHD} and \emph{ID}
In case of \emph{ida} and \emph{ga} the order of the parameters is: \emph{kg}, \emph{I0}, \emph{IHD} and \emph{ID}.}

\item{seed}{is an optional integer argument defining the seed which is set directly for the optimization. In case the argument is not set the current time is used as seed.}

\item{npop}{is an optional integer argument defining the number of particles during optimization. The default value is set to 40.}

\item{ngen}{is an optional integer argument defining the number of generations of the particle swarm optimization. The default value is set to 200.}

\item{Topology}{is an optional character argument defining which topology should be used by the particle swarm algorithm. The options are "star" and "random". The default topology is the "random" topology.}

\item{errorThreshold}{is an optional numeric argument defining a sufficient small error which acts as a stop signal for the particle swarm algorithm. The default value is set to -Inf.}

\item{num_rep}{is an optional integer argument defining the number of replicates for each dataset}

\item{num_cores}{is an optional integer argument defining the maximum number of cores which should be used for the optimization}
}
\description{
Runs a batch of optimization tasks
}
\examples{
path <- paste0(system.file("examples", package = "tsf"), "/IDABatch.csv")
lowerBounds <- c(
  kG = 1000,
  I0 = 0,
  IHD = 0,
  ID = 0
)
upperBounds <- c(
  kG = 10^8,
  I0 = 100,
  IHD = 10^7,
  ID = 10^7
)
additionalParameters <- c(
  host = 1e-6,
  dye = 1e-6,
  kHD = 3e6
)
tsf::batch(
  "ida",
  lowerBounds, upperBounds,
  path, additionalParameters,
  ngen = 20
)
}
