% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateLossFunction.R
\name{createLossFunction}
\alias{createLossFunction}
\title{Create a costume loss function}
\usage{
createLossFunction(
  f,
  listElimVars,
  unknownRoots,
  parameterVec,
  additionalParameterVec,
  variedParameterVec,
  linearSystem
)
}
\arguments{
\item{f}{is a function defining the algebraic system}

\item{listElimVars}{a list with one or more character vectors which define the variables which have to be eliminated respectively}

\item{unknownRoots}{a character vector defining the names of the unknown variables.}

\item{parameterVec}{a character vector defining the parameters which have to be optimized}

\item{additionalParameterVec}{a character vector defining additional parameters which are constant.}

\item{variedParameterVec}{a character vector with two entries defining the name of the variable which is altered, for instance d0 when dye is added to the system. The second entry has to be the independent variable e.g. signal}

\item{linearSystem}{a function defining the linear system which is used to calculate the insilico dependent variable}
}
\value{
the resulting loss function
}
\description{
Create a costume loss function
}
\examples{
f <- function() {
  h + hd + -h0 = 0
  d + hd -d0 = 0   
  hd / (h*d) -kd = 0
}
listElimVars <- list(c("h", "d"), c("hd", "h"))
unknownRoots <- c("d", "hd")
parameterVec <- c("kd", "I0", "IHD", "ID")
additionalParameterVec <- c("h0")
variedParameterVec <- c("d0", "signal")
linearSystem <- function() {I0 + IHD * hd + ID * d}
createLossFunction(f, listElimVars, unknownRoots, parameterVec, additionalParameterVec,
              variedParameterVec, linearSystem)
}
