% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Optimize algebraic systems which describe thermodynamic binding systems}
\usage{
runApp()
}
\description{
Optimize algebraic systems which describe thermodynamic binding systems
}
