% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterSensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Optimize algebraic systems which describe thermodynamic binding systems}
\usage{
sensitivity(
  case,
  parameters,
  path,
  additionalParameters,
  percentage = NULL,
  OffsetBoundaries = NULL
)
}
\arguments{
\item{case}{is a character describing which system should be investigated. Either: "hg", "ida" or "gda".}

\item{parameters}{is a numeric vector containing already optimized parameter.
In case of \emph{hg} the order of the parameters is: \emph{khd}, \emph{I0}, \emph{IHD} and \emph{ID}
In case of \emph{ida} and \emph{ga} the order of the parameters is: \emph{kg}, \emph{I0}, \emph{IHD} and \emph{ID}.}

\item{path}{is a filepath which contains tabular x-y data. The concentraion of dye or guest respectivly is assumed to be in the first colum. Furthermore, should the corresponding signal be stored in the second column.}

\item{additionalParameters}{are required parameters which are specific for each case.
In case of \emph{hg} a numeric vector of length 1 is expected which contains the concentration of the host.
In case of \emph{ida} a numeric vector of length 3 is expected which contains the concentration of the host, dye and the \emph{khd} parameter.
In case of \emph{gda} a numeric vector of length 3 is expected which contains the concentration of the host, guest and the \emph{khd} parameter.}

\item{percentage}{is the percentage +/- from parameters in which the sensitivity should be analysed.}

\item{OffsetBoundaries}{in case percentage is not suitable a numeric vector (equivalent to parameters) can be used which is added/substracted from parameters. It is only possible to set either percentage or OffsetBoundaries.}
}
\value{
either an instance of ErrorClass if something went wrong. Otherwise plots showing the sensitivity are returned.
}
\description{
Optimize algebraic systems which describe thermodynamic binding systems
}
\examples{
path <- paste0(system.file("examples", package = "tsf"), "/IDA.txt")
res <- opti("ida", c(1, 0, 0, 0), c(10^9, 10^6, 10^6, 10^6), path, c(4.3, 6.0, 7079458)) 
sensitivity("ida", res[[2]], path, c(4.3, 6.0, 7079458), 20)
}
